Update qQCReport.Rd
This is a quick fix following the question on the BioC mailing list (https://support.bioconductor.org/p/118580/). It is still not correct, e.g. what happens when you generate a qProject with bam files which do not contain the quality.....

Hence, it is rather work in progress and should not be merged yet, but might trigger a discussion.